frontend-book/vercel/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
qdrant-client==1.8.0
cohere==4.44
python-dotenv==1.0.0
trafilatura==1.6.1
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
openai==1.12.0
httpx==0.25.2
tiktoken==0.8.0
